create motor app
