+++ RaspberryPI/blubblub-webserver/manage.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python
 import os
 import sys
-from sensor import sensors
 
 if __name__ == "__main__":
     os.environ.setdefault("DJANGO_SETTINGS_MODULE", "blubblub.settings")
@@ -21,6 +20,4 @@ if __name__ == "__main__":
             )
         raise
 
-    s = sensors.Sensors()
-    s.start()
     execute_from_command_line(sys.argv)
