+++ RaspberryPI/blubblub-webserver/motor/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class MotorConfig(AppConfig):
+    name = 'motor'
