+++ RaspberryPI/blubblub-webserver/motor/urls.py
@@ -2,5 +2,5 @@ from django.conf.urls import url
 from . import views
 
 urlpatterns = [
-	url(r'^sensor/$', views.showData, name='showData'),
+	url(r'^feeding/$', views.feeding, name='feeding'),
 ]
